split celtest tests by behavior
Move celtest coverage out of the monolithic evaluator test file into focused
test files for admission evaluation, mutation evaluation, match conditions,
audit annotations, parsing, and sample-style parse-and-evaluate cases.

Keep evaluator_test focused on evaluator construction, compile checks,
expression evaluation, version behavior, cost behavior, and violation
formatting.
