+++ staging/src/k8s.io/apiserver/pkg/admission/testing/celtest/evaluator_test.go
@@ -18,14 +18,8 @@ package celtest
 
 import (
 	"fmt"
-	"os"
-	"path/filepath"
 	"strings"
 	"testing"
-
-	admissionv1 "k8s.io/api/admission/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 )
 
 func TestNewEvaluator(t *testing.T) {
@@ -151,949 +145,6 @@ func TestEvalExpression(t *testing.T) {
 	}
 }
 
-func TestEvalAdmission_Allowed(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{Path: "validations[0]", Expression: "object.metadata.name == 'allowed-name'"},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "allowed-name"},
-		},
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if !result.Allowed {
-		t.Errorf("EvalAdmission() Allowed = false, want true; violations: %s", result.FormatViolations())
-	}
-	if result.Cost <= 0 {
-		t.Errorf("EvalAdmission() Cost = %d, want > 0", result.Cost)
-	}
-}
-
-func TestEvalAdmission_Denied(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{
-				Path:       "validations[0]",
-				Expression: "object.metadata.name != 'bad-name'",
-				Message:    "name must not be bad-name",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "bad-name"},
-		},
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if result.Allowed {
-		t.Error("EvalAdmission() Allowed = true, want false")
-	}
-	if len(result.Violations) != 1 {
-		t.Fatalf("EvalAdmission() got %d violations, want 1", len(result.Violations))
-	}
-	if result.Violations[0].Message != "name must not be bad-name" {
-		t.Errorf("violation message = %q, want %q", result.Violations[0].Message, "name must not be bad-name")
-	}
-}
-
-func TestEvalAdmission_MessageExpression(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{
-				Path:              "validations[0]",
-				Expression:        "false",
-				Message:           "static fallback",
-				MessageExpression: "'denied: ' + object.metadata.name",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "my-pod"},
-		},
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if result.Allowed {
-		t.Error("EvalAdmission() Allowed = true, want false")
-	}
-	if len(result.Violations) != 1 {
-		t.Fatalf("got %d violations, want 1", len(result.Violations))
-	}
-	if result.Violations[0].Message != "denied: my-pod" {
-		t.Errorf("violation message = %q, want %q", result.Violations[0].Message, "denied: my-pod")
-	}
-	if result.Violations[0].MessageError != nil {
-		t.Errorf("unexpected MessageError: %v", result.Violations[0].MessageError)
-	}
-}
-
-func TestEvalAdmission_MessageExpressionError(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{
-				Path:              "validations[0]",
-				Expression:        "false",
-				Message:           "static fallback",
-				MessageExpression: "object.nonexistent.deeply.nested",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test"},
-		},
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if len(result.Violations) != 1 {
-		t.Fatalf("got %d violations, want 1", len(result.Violations))
-	}
-	if result.Violations[0].MessageError == nil {
-		t.Error("expected MessageError to be non-nil for invalid messageExpression")
-	}
-	// Static message should still be populated as fallback
-	if result.Violations[0].Message != "static fallback" {
-		t.Errorf("violation message = %q, want %q", result.Violations[0].Message, "static fallback")
-	}
-}
-
-func TestEvalAdmission_MessageExpressionDoesNotDeclareAuthorizer(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{
-				Path:              "validations[0]",
-				Expression:        "false",
-				Message:           "static fallback",
-				MessageExpression: "authorizer.requestResource.check('get').allowed() ? 'allowed' : 'denied'",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test"},
-		},
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if len(result.Violations) != 1 {
-		t.Fatalf("got %d violations, want 1", len(result.Violations))
-	}
-	if result.Violations[0].MessageError == nil {
-		t.Fatal("expected MessageError when messageExpression references authorizer")
-	}
-	if result.Violations[0].Message != "static fallback" {
-		t.Errorf("violation message = %q, want fallback", result.Violations[0].Message)
-	}
-
-	_, err = e.EvalValidation(policy, ValidationSelector{Path: "validations[0]", Part: "messageExpression"}, input)
-	if err == nil {
-		t.Fatal("expected EvalValidation messageExpression to reject authorizer")
-	}
-}
-
-func TestEvalAdmission_Variables(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Variables: []Variable{
-			{Name: "podName", Expression: "object.metadata.name"},
-			{Name: "isAllowed", Expression: "variables.podName == 'good-pod'"},
-		},
-		Validations: []Validation{
-			{Path: "validations[0]", Expression: "variables.isAllowed"},
-		},
-	}
-
-	t.Run("allowed", func(t *testing.T) {
-		input := &AdmissionInput{
-			Object: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "good-pod"},
-			},
-		}
-		result, err := e.EvalAdmission(policy, input)
-		if err != nil {
-			t.Fatalf("EvalAdmission() error: %v", err)
-		}
-		if !result.Allowed {
-			t.Errorf("expected Allowed=true, got violations: %s", result.FormatViolations())
-		}
-	})
-
-	t.Run("denied", func(t *testing.T) {
-		input := &AdmissionInput{
-			Object: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "bad-pod"},
-			},
-		}
-		result, err := e.EvalAdmission(policy, input)
-		if err != nil {
-			t.Fatalf("EvalAdmission() error: %v", err)
-		}
-		if result.Allowed {
-			t.Error("expected Allowed=false")
-		}
-	})
-}
-
-func TestEvalVariable(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Variables: []Variable{
-			{Name: "podName", Expression: "object.metadata.name"},
-			{Name: "nameLen", Expression: "size(variables.podName)"},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "hello"},
-		},
-	}
-
-	t.Run("first variable", func(t *testing.T) {
-		val, err := e.EvalVariable(policy, "podName", input)
-		if err != nil {
-			t.Fatalf("EvalVariable() error: %v", err)
-		}
-		if val != "hello" {
-			t.Errorf("EvalVariable() = %v, want %q", val, "hello")
-		}
-	})
-
-	t.Run("chained variable", func(t *testing.T) {
-		val, err := e.EvalVariable(policy, "nameLen", input)
-		if err != nil {
-			t.Fatalf("EvalVariable() error: %v", err)
-		}
-		if val != int64(5) {
-			t.Errorf("EvalVariable() = %v (%T), want 5 (int64)", val, val)
-		}
-	})
-
-	t.Run("not found", func(t *testing.T) {
-		_, err := e.EvalVariable(policy, "nonexistent", input)
-		if err == nil {
-			t.Error("EvalVariable() expected error for missing variable")
-		}
-	})
-}
-
-func TestEvalValidation(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{
-				Path:              "validations[0]",
-				Expression:        "object.metadata.name == 'test'",
-				Message:           "name must be test",
-				MessageExpression: "'got: ' + object.metadata.name",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test"},
-		},
-	}
-
-	t.Run("expression part", func(t *testing.T) {
-		val, err := e.EvalValidation(policy, ValidationSelector{Path: "validations[0]"}, input)
-		if err != nil {
-			t.Fatalf("EvalValidation() error: %v", err)
-		}
-		if val != true {
-			t.Errorf("EvalValidation() = %v, want true", val)
-		}
-	})
-
-	t.Run("messageExpression part", func(t *testing.T) {
-		val, err := e.EvalValidation(policy, ValidationSelector{Path: "validations[0]", Part: "messageExpression"}, input)
-		if err != nil {
-			t.Fatalf("EvalValidation() error: %v", err)
-		}
-		if val != "got: test" {
-			t.Errorf("EvalValidation() = %v, want %q", val, "got: test")
-		}
-	})
-
-	t.Run("empty path", func(t *testing.T) {
-		_, err := e.EvalValidation(policy, ValidationSelector{}, input)
-		if err == nil {
-			t.Error("expected error for empty path")
-		}
-	})
-
-	t.Run("not found path", func(t *testing.T) {
-		_, err := e.EvalValidation(policy, ValidationSelector{Path: "validations[99]"}, input)
-		if err == nil {
-			t.Error("expected error for missing path")
-		}
-	})
-
-	t.Run("unsupported part", func(t *testing.T) {
-		_, err := e.EvalValidation(policy, ValidationSelector{Path: "validations[0]", Part: "bogus"}, input)
-		if err == nil {
-			t.Error("expected error for unsupported part")
-		}
-	})
-}
-
-func TestEvalMatchConditions(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata": map[string]interface{}{
-				"name":      "test-pod",
-				"namespace": "default",
-			},
-		},
-		Params: map[string]interface{}{
-			"data": map[string]interface{}{"enabled": "true"},
-		},
-	}
-
-	t.Run("all match", func(t *testing.T) {
-		policy := &AdmissionPolicy{
-			MatchConditions: []MatchCondition{
-				{Path: "spec.matchConditions[0]", Name: "params-enabled", Expression: "params.data.enabled == 'true'"},
-				{Path: "spec.matchConditions[1]", Name: "not-system", Expression: "object.metadata.namespace != 'kube-system'"},
-			},
-		}
-		policy.SetHasParams(true)
-
-		result, err := e.EvalMatchConditions(policy, input)
-		if err != nil {
-			t.Fatalf("EvalMatchConditions() error: %v", err)
-		}
-		if len(result.Conditions) != 2 {
-			t.Fatalf("got %d conditions, want 2", len(result.Conditions))
-		}
-		for index, condition := range result.Conditions {
-			if condition.Error != nil {
-				t.Fatalf("condition[%d] error: %v", index, condition.Error)
-			}
-			if condition.Value != true {
-				t.Fatalf("condition[%d] value = %v, want true", index, condition.Value)
-			}
-		}
-	})
-
-	t.Run("false condition", func(t *testing.T) {
-		policy := &AdmissionPolicy{
-			MatchConditions: []MatchCondition{
-				{Path: "spec.matchConditions[0]", Name: "system-only", Expression: "object.metadata.namespace == 'kube-system'"},
-			},
-		}
-
-		result, err := e.EvalMatchConditions(policy, input)
-		if err != nil {
-			t.Fatalf("EvalMatchConditions() error: %v", err)
-		}
-		if len(result.Conditions) != 1 {
-			t.Fatalf("got %d conditions, want 1", len(result.Conditions))
-		}
-		if result.Conditions[0].Error != nil {
-			t.Fatalf("condition error: %v", result.Conditions[0].Error)
-		}
-		if result.Conditions[0].Value != false {
-			t.Fatalf("condition value = %v, want false", result.Conditions[0].Value)
-		}
-	})
-
-	t.Run("runtime error is reported per condition", func(t *testing.T) {
-		policy := &AdmissionPolicy{
-			MatchConditions: []MatchCondition{
-				{Path: "spec.matchConditions[0]", Name: "bad", Expression: "1 / 0 == 0"},
-			},
-		}
-
-		result, err := e.EvalMatchConditions(policy, input)
-		if err != nil {
-			t.Fatalf("EvalMatchConditions() error: %v", err)
-		}
-		if len(result.Conditions) != 1 || result.Conditions[0].Error == nil {
-			t.Fatalf("expected per-condition error, got %#v", result.Conditions)
-		}
-	})
-
-	t.Run("compile error", func(t *testing.T) {
-		policy := &AdmissionPolicy{
-			MatchConditions: []MatchCondition{
-				{Path: "spec.matchConditions[0]", Name: "bad", Expression: "object.metadata.name =="},
-			},
-		}
-
-		if _, err := e.EvalMatchConditions(policy, input); err == nil {
-			t.Fatal("expected compile error")
-		}
-	})
-
-	t.Run("single condition selector by name", func(t *testing.T) {
-		policy := &AdmissionPolicy{
-			MatchConditions: []MatchCondition{
-				{Path: "spec.matchConditions[0]", Name: "params-enabled", Expression: "params.data.enabled == 'true'"},
-			},
-		}
-		policy.SetHasParams(true)
-
-		value, err := e.EvalMatchCondition(policy, MatchConditionSelector{Name: "params-enabled"}, input)
-		if err != nil {
-			t.Fatalf("EvalMatchCondition() error: %v", err)
-		}
-		if value != true {
-			t.Errorf("EvalMatchCondition() = %v, want true", value)
-		}
-	})
-
-	t.Run("duplicate names require path", func(t *testing.T) {
-		policy := &AdmissionPolicy{
-			MatchConditions: []MatchCondition{
-				{Path: "webhooks[0].matchConditions[0]", Name: "same-name", Expression: "false"},
-				{Path: "webhooks[1].matchConditions[0]", Name: "same-name", Expression: "true"},
-			},
-		}
-
-		if _, err := e.EvalMatchCondition(policy, MatchConditionSelector{Name: "same-name"}, input); err == nil {
-			t.Fatal("expected ambiguous selector error")
-		}
-
-		value, err := e.EvalMatchCondition(policy, MatchConditionSelector{Path: "webhooks[1].matchConditions[0]"}, input)
-		if err != nil {
-			t.Fatalf("EvalMatchCondition() error: %v", err)
-		}
-		if value != true {
-			t.Errorf("EvalMatchCondition() = %v, want true", value)
-		}
-	})
-}
-
-func TestEvalAdmission_EvaluatesValidationsIndependentlyOfMatchConditions(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		MatchConditions: []MatchCondition{{Path: "spec.matchConditions[0]", Name: "not-system", Expression: "false"}},
-		Validations:     []Validation{{Path: "spec.validations[0]", Expression: "false", Message: "validation still evaluated"}},
-	}
-
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata": map[string]interface{}{
-				"name":      "test-pod",
-				"namespace": "kube-system",
-			},
-		},
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if result.Allowed {
-		t.Fatal("expected validation to be evaluated and deny")
-	}
-	if len(result.Violations) != 1 {
-		t.Fatalf("got %d violations, want 1", len(result.Violations))
-	}
-}
-
-func TestEvalMutation_EvaluatesMutationsIndependentlyOfMatchConditions(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		MatchConditions: []MatchCondition{{Path: "spec.matchConditions[0]", Name: "not-system", Expression: "false"}},
-		Mutations: []Mutation{{
-			Path:       "spec.mutations[0]",
-			PatchType:  "ApplyConfiguration",
-			Expression: "Object{metadata: Object.metadata{labels: {'mutated': 'true'}}}",
-		}},
-	}
-
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata": map[string]interface{}{
-				"name":      "test-pod",
-				"namespace": "kube-system",
-			},
-		},
-	}
-
-	result, err := e.EvalMutation(policy, input)
-	if err != nil {
-		t.Fatalf("EvalMutation() error: %v", err)
-	}
-	if len(result.Patches) != 1 {
-		t.Fatalf("got %d patches, want 1", len(result.Patches))
-	}
-}
-
-func TestParseAndEvalVAPMatchConditionsAndAuditAnnotations(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policyYAML := `{"apiVersion":"admissionregistration.k8s.io/v1","kind":"ValidatingAdmissionPolicy","spec":{"paramKind":{"apiVersion":"v1","kind":"ConfigMap"},"matchConditions":[{"name":"enabled","expression":"params.data.enabled == 'true'"}],"validations":[{"expression":"true"}],"auditAnnotations":[{"key":"pod-name","valueExpression":"string(object.metadata.name)"}]}}`
-	policy, err := ParseAdmissionPolicy(policyYAML)
-	if err != nil {
-		t.Fatalf("ParseAdmissionPolicy() error: %v", err)
-	}
-
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test-pod"},
-		},
-		Params: map[string]interface{}{
-			"data": map[string]interface{}{"enabled": "true"},
-		},
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if !result.Allowed {
-		t.Fatalf("expected admission validations to allow: %s", result.FormatViolations())
-	}
-
-	matchResult, err := e.EvalMatchConditions(policy, input)
-	if err != nil {
-		t.Fatalf("EvalMatchConditions() error: %v", err)
-	}
-	if len(matchResult.Conditions) != 1 || matchResult.Conditions[0].Value != true {
-		t.Fatalf("unexpected matchConditions: %#v", matchResult.Conditions)
-	}
-
-	auditResult, err := e.EvalAuditAnnotations(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAuditAnnotations() error: %v", err)
-	}
-	if len(auditResult.Annotations) != 1 || auditResult.Annotations[0].Value != "test-pod" {
-		t.Fatalf("unexpected audit annotations: %#v", auditResult.Annotations)
-	}
-}
-
-func TestParseAndEvalMAPMatchConditions(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policyYAML := `{"apiVersion":"admissionregistration.k8s.io/v1","kind":"MutatingAdmissionPolicy","spec":{"matchConditions":[{"name":"only-pods","expression":"object.kind == 'Pod'"}],"mutations":[{"patchType":"ApplyConfiguration","applyConfiguration":{"expression":"Object{metadata: Object.metadata{labels: {'mutated': 'true'}}}"}}]}}`
-	policy, err := ParseAdmissionPolicy(policyYAML)
-	if err != nil {
-		t.Fatalf("ParseAdmissionPolicy() error: %v", err)
-	}
-
-	t.Run("false match condition still allows direct mutation evaluation", func(t *testing.T) {
-		input := &AdmissionInput{
-			Object: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "ConfigMap",
-				"metadata":   map[string]interface{}{"name": "not-a-pod"},
-			},
-		}
-		matchResult, err := e.EvalMatchConditions(policy, input)
-		if err != nil {
-			t.Fatalf("EvalMatchConditions() error: %v", err)
-		}
-		if len(matchResult.Conditions) != 1 || matchResult.Conditions[0].Value != false {
-			t.Fatalf("unexpected matchConditions: %#v", matchResult.Conditions)
-		}
-
-		result, err := e.EvalMutation(policy, input)
-		if err != nil {
-			t.Fatalf("EvalMutation() error: %v", err)
-		}
-		if len(result.Patches) != 1 {
-			t.Fatalf("got %d patches, want 1", len(result.Patches))
-		}
-	})
-
-	t.Run("true match condition", func(t *testing.T) {
-		input := &AdmissionInput{
-			Object: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "pod"},
-			},
-		}
-		matchResult, err := e.EvalMatchConditions(policy, input)
-		if err != nil {
-			t.Fatalf("EvalMatchConditions() error: %v", err)
-		}
-		if len(matchResult.Conditions) != 1 || matchResult.Conditions[0].Value != true {
-			t.Fatalf("unexpected matchConditions: %#v", matchResult.Conditions)
-		}
-
-		result, err := e.EvalMutation(policy, input)
-		if err != nil {
-			t.Fatalf("EvalMutation() error: %v", err)
-		}
-		if len(result.Patches) != 1 {
-			t.Fatalf("got %d patches, want 1", len(result.Patches))
-		}
-	})
-}
-
-func TestEvalAuditAnnotations(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		AuditAnnotations: []AuditAnnotation{
-			{Path: "spec.auditAnnotations[0]", Key: "pod-name", ValueExpression: "string(object.metadata.name)"},
-			{Path: "spec.auditAnnotations[1]", Key: "empty", ValueExpression: "''"},
-			{Path: "spec.auditAnnotations[2]", Key: "null", ValueExpression: "null"},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "audit-pod"},
-		},
-	}
-
-	result, err := e.EvalAuditAnnotations(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAuditAnnotations() error: %v", err)
-	}
-	if len(result.Annotations) != 3 {
-		t.Fatalf("got %d annotations, want 3", len(result.Annotations))
-	}
-	if result.Annotations[0].Value != "audit-pod" {
-		t.Errorf("annotation[0] = %#v, want audit-pod", result.Annotations[0])
-	}
-	if result.Annotations[1].Value != "" {
-		t.Errorf("annotation[1] value = %v, want empty string", result.Annotations[1].Value)
-	}
-	if result.Annotations[2].Value != nil {
-		t.Errorf("annotation[2] value = %v, want nil", result.Annotations[2].Value)
-	}
-
-	value, err := e.EvalAuditAnnotation(policy, AuditAnnotationSelector{Path: "spec.auditAnnotations[0]"}, input)
-	if err != nil {
-		t.Fatalf("EvalAuditAnnotation() error: %v", err)
-	}
-	if value != "audit-pod" {
-		t.Errorf("EvalAuditAnnotation() = %v, want audit-pod", value)
-	}
-}
-
-func TestEvalAuditAnnotations_CompileError(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test-pod"},
-		},
-	}
-
-	policy := &AdmissionPolicy{
-		Validations:      []Validation{{Path: "spec.validations[0]", Expression: "true"}},
-		AuditAnnotations: []AuditAnnotation{{Path: "spec.auditAnnotations[0]", Key: "bad", ValueExpression: "1"}},
-	}
-
-	if _, err := e.EvalAuditAnnotations(policy, input); err == nil {
-		t.Fatal("expected audit annotation compile error")
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if !result.Allowed {
-		t.Fatalf("EvalAdmission should only evaluate validations, got violations: %s", result.FormatViolations())
-	}
-}
-
-func TestEvalAdmission_RequestFields(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{Path: "validations[0]", Expression: "request.operation == 'CREATE'"},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test"},
-		},
-		Request: &admissionv1.AdmissionRequest{
-			Operation: admissionv1.Create,
-			Kind:      metav1.GroupVersionKind{Version: "v1", Kind: "Pod"},
-			Resource:  metav1.GroupVersionResource{Version: "v1", Resource: "pods"},
-		},
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if !result.Allowed {
-		t.Errorf("expected Allowed=true, got violations: %s", result.FormatViolations())
-	}
-}
-
-func TestParseAdmissionPolicy_Flat(t *testing.T) {
-	yaml := `
-variables:
-  - name: podName
-    expression: "object.metadata.name"
-validations:
-  - expression: "variables.podName != 'bad'"
-    message: "name must not be bad"
-`
-	policy, err := ParseAdmissionPolicy(yaml)
-	if err != nil {
-		t.Fatalf("ParseAdmissionPolicy() error: %v", err)
-	}
-	if len(policy.Variables) != 1 {
-		t.Errorf("got %d variables, want 1", len(policy.Variables))
-	}
-	if len(policy.Validations) != 1 {
-		t.Errorf("got %d validations, want 1", len(policy.Validations))
-	}
-	if policy.Variables[0].Name != "podName" {
-		t.Errorf("variable name = %q, want %q", policy.Variables[0].Name, "podName")
-	}
-}
-
-func TestParseAdmissionPolicy_VAP(t *testing.T) {
-	yaml := `
-apiVersion: admissionregistration.k8s.io/v1
-kind: ValidatingAdmissionPolicy
-spec:
-  matchConditions:
-    - name: skip-system-namespaces
-      expression: "request.namespace != 'kube-system'"
-  validations:
-    - expression: "object.spec.replicas <= 5"
-      message: "too many replicas"
-  auditAnnotations:
-    - key: replicas
-      valueExpression: "string(object.spec.replicas)"
-`
-	policy, err := ParseAdmissionPolicy(yaml)
-	if err != nil {
-		t.Fatalf("ParseAdmissionPolicy() error: %v", err)
-	}
-	if len(policy.MatchConditions) != 1 {
-		t.Fatalf("got %d matchConditions, want 1", len(policy.MatchConditions))
-	}
-	if policy.MatchConditions[0].Path != "spec.matchConditions[0]" {
-		t.Errorf("matchCondition[0] path = %q, want spec.matchConditions[0]", policy.MatchConditions[0].Path)
-	}
-	if len(policy.Validations) != 1 {
-		t.Errorf("got %d validations, want 1", len(policy.Validations))
-	}
-	if len(policy.AuditAnnotations) != 1 {
-		t.Errorf("got %d auditAnnotations, want 1", len(policy.AuditAnnotations))
-	}
-	if policy.hasParams {
-		t.Error("hasParams should be false when paramKind is not set")
-	}
-}
-
-func TestParseAdmissionPolicy_Errors(t *testing.T) {
-	tests := []struct {
-		name string
-		yaml string
-	}{
-		{name: "empty variable name", yaml: `
-variables:
-  - name: ""
-    expression: "true"
-`},
-		{name: "empty variable expression", yaml: `
-variables:
-  - name: "x"
-    expression: ""
-`},
-		{name: "empty validation expression", yaml: `
-validations:
-  - expression: ""
-`},
-		{name: "unsupported kind", yaml: `
-apiVersion: v1
-kind: UnsupportedKind
-`},
-		{name: "empty policy", yaml: `{}`},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, err := ParseAdmissionPolicy(tt.yaml)
-			if err == nil {
-				t.Error("ParseAdmissionPolicy() expected error")
-			}
-		})
-	}
-}
-
-func TestEvalAdmission_CompilationError(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{Path: "validations[0]", Expression: "???invalid"},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test"},
-		},
-	}
-
-	_, err = e.EvalAdmission(policy, input)
-	if err == nil {
-		t.Error("EvalAdmission() expected compilation error")
-	}
-}
-
-func TestEvalAdmission_MultipleValidations(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{Path: "validations[0]", Expression: "true"},
-			{Path: "validations[1]", Expression: "false", Message: "always fails"},
-			{Path: "validations[2]", Expression: "true"},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test"},
-		},
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if result.Allowed {
-		t.Error("expected Allowed=false when one validation fails")
-	}
-	if len(result.Violations) != 1 {
-		t.Fatalf("got %d violations, want 1", len(result.Violations))
-	}
-	if result.Violations[0].Message != "always fails" {
-		t.Errorf("violation message = %q, want %q", result.Violations[0].Message, "always fails")
-	}
-}
-
 func TestFormatViolations(t *testing.T) {
 	t.Run("no violations", func(t *testing.T) {
 		r := &AdmissionResult{Allowed: true}
@@ -1118,523 +169,6 @@ func TestFormatViolations(t *testing.T) {
 		got := r.FormatViolations()
 		if !strings.Contains(got, "compile error") {
 			t.Errorf("FormatViolations() = %q, expected to contain 'compile error'", got)
-		}
-	})
-}
-
-func TestPreambleVariables(t *testing.T) {
-	e, err := NewEvaluator(
-		WithPreambleVariables(
-			Variable{Name: "alwaysTrue", Expression: "true"},
-		),
-	)
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{Path: "validations[0]", Expression: "variables.alwaysTrue"},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test"},
-		},
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if !result.Allowed {
-		t.Errorf("expected Allowed=true with preamble variable, got violations: %s", result.FormatViolations())
-	}
-}
-
-func TestParseAdmissionPolicy_MAP(t *testing.T) {
-	yaml := `{"apiVersion":"admissionregistration.k8s.io/v1","kind":"MutatingAdmissionPolicy","spec":{"matchConditions":[{"name":"only-pods","expression":"request.resource.resource == 'pods'"}],"variables":[{"name":"replicas","expression":"object.spec.replicas"}],"mutations":[{"patchType":"ApplyConfiguration","applyConfiguration":{"expression":"Object{spec: Object.spec{replicas: 3}}"}},{"patchType":"JSONPatch","jsonPatch":{"expression":"[JSONPatch{op: \"replace\", path: \"/spec/replicas\", value: 3}]"}}]}}`
-	policy, err := ParseAdmissionPolicy(yaml)
-	if err != nil {
-		t.Fatalf("ParseAdmissionPolicy() error: %v", err)
-	}
-	if len(policy.Variables) != 1 {
-		t.Errorf("got %d variables, want 1", len(policy.Variables))
-	}
-	if len(policy.MatchConditions) != 1 {
-		t.Errorf("got %d matchConditions, want 1", len(policy.MatchConditions))
-	}
-	if len(policy.Mutations) != 2 {
-		t.Fatalf("got %d mutations, want 2", len(policy.Mutations))
-	}
-	if policy.Mutations[0].PatchType != "ApplyConfiguration" {
-		t.Errorf("mutation[0] patchType = %q, want ApplyConfiguration", policy.Mutations[0].PatchType)
-	}
-	if policy.Mutations[1].PatchType != "JSONPatch" {
-		t.Errorf("mutation[1] patchType = %q, want JSONPatch", policy.Mutations[1].PatchType)
-	}
-	if policy.hasParams {
-		t.Error("hasParams should be false when paramKind is not set")
-	}
-}
-
-func TestParseAdmissionPolicy_MAP_WithParamKind(t *testing.T) {
-	yaml := `
-apiVersion: admissionregistration.k8s.io/v1
-kind: MutatingAdmissionPolicy
-spec:
-  paramKind:
-    apiVersion: v1
-    kind: ConfigMap
-  mutations:
-    - patchType: ApplyConfiguration
-      applyConfiguration:
-        expression: "Object{spec: Object.spec{replicas: int(params.data.maxReplicas)}}"
-`
-	policy, err := ParseAdmissionPolicy(yaml)
-	if err != nil {
-		t.Fatalf("ParseAdmissionPolicy() error: %v", err)
-	}
-	if !policy.hasParams {
-		t.Error("hasParams should be true when paramKind is set")
-	}
-}
-
-func TestParseAdmissionPolicy_MAP_Errors(t *testing.T) {
-	tests := []struct {
-		name string
-		yaml string
-	}{
-		{name: "missing patchType", yaml: `
-apiVersion: admissionregistration.k8s.io/v1
-kind: MutatingAdmissionPolicy
-spec:
-  mutations:
-    - applyConfiguration:
-        expression: "Object{}"
-`},
-		{name: "missing applyConfiguration expression", yaml: `
-apiVersion: admissionregistration.k8s.io/v1
-kind: MutatingAdmissionPolicy
-spec:
-  mutations:
-    - patchType: ApplyConfiguration
-`},
-		{name: "missing jsonPatch expression", yaml: `
-apiVersion: admissionregistration.k8s.io/v1
-kind: MutatingAdmissionPolicy
-spec:
-  mutations:
-    - patchType: JSONPatch
-`},
-		{name: "unsupported patchType", yaml: `
-apiVersion: admissionregistration.k8s.io/v1
-kind: MutatingAdmissionPolicy
-spec:
-  mutations:
-    - patchType: StrategicMerge
-      applyConfiguration:
-        expression: "Object{}"
-`},
-		{name: "empty policy", yaml: `
-apiVersion: admissionregistration.k8s.io/v1
-kind: MutatingAdmissionPolicy
-spec: {}
-`},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, err := ParseAdmissionPolicy(tt.yaml)
-			if err == nil {
-				t.Error("ParseAdmissionPolicy() expected error")
-			}
-		})
-	}
-}
-
-func TestEvalMutation_ApplyConfiguration(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Variables: []Variable{
-			{Name: "newName", Expression: "'mutated-pod'"},
-		},
-		Mutations: []Mutation{
-			{
-				Path:       "spec.mutations[0]",
-				PatchType:  "ApplyConfiguration",
-				Expression: "Object{metadata: Object.metadata{name: variables.newName}}",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "original-pod"},
-		},
-	}
-
-	result, err := e.EvalMutation(policy, input)
-	if err != nil {
-		t.Fatalf("EvalMutation() error: %v", err)
-	}
-	if len(result.Patches) != 1 {
-		t.Fatalf("got %d patches, want 1", len(result.Patches))
-	}
-	if result.Patches[0].Error != nil {
-		t.Fatalf("patch error: %v", result.Patches[0].Error)
-	}
-	if result.Patches[0].PatchType != "ApplyConfiguration" {
-		t.Errorf("patchType = %q, want ApplyConfiguration", result.Patches[0].PatchType)
-	}
-	applyConfig, ok := result.Patches[0].Value.(map[string]interface{})
-	if !ok {
-		t.Fatalf("patch value type = %T, want map[string]interface{}", result.Patches[0].Value)
-	}
-	metadata, ok := applyConfig["metadata"].(map[string]interface{})
-	if !ok {
-		t.Fatalf("metadata type = %T, want map[string]interface{}", applyConfig["metadata"])
-	}
-	if metadata["name"] != "mutated-pod" {
-		t.Errorf("metadata.name = %v, want %q", metadata["name"], "mutated-pod")
-	}
-	if result.Cost <= 0 {
-		t.Errorf("Cost = %d, want > 0", result.Cost)
-	}
-}
-
-func TestEvalMutation_JSONPatch(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Mutations: []Mutation{
-			{
-				Path:       "spec.mutations[0]",
-				PatchType:  "JSONPatch",
-				Expression: "[JSONPatch{op: \"replace\", path: \"/metadata/name\", value: \"new-name\"}]",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "old-name"},
-		},
-	}
-
-	result, err := e.EvalMutation(policy, input)
-	if err != nil {
-		t.Fatalf("EvalMutation() error: %v", err)
-	}
-	if len(result.Patches) != 1 {
-		t.Fatalf("got %d patches, want 1", len(result.Patches))
-	}
-	if result.Patches[0].Error != nil {
-		t.Fatalf("patch error: %v", result.Patches[0].Error)
-	}
-	// JSONPatch expressions return a CEL list of JSONPatch values;
-	// verify the result is a non-empty slice.
-	patchOps, ok := result.Patches[0].Value.([]interface{})
-	if !ok {
-		// CEL may return []ref.Val for typed JSONPatch objects; accept any non-nil slice.
-		if result.Patches[0].Value == nil {
-			t.Fatal("patch value should not be nil")
-		}
-	} else if len(patchOps) != 1 {
-		t.Errorf("got %d JSON patch ops, want 1", len(patchOps))
-	}
-}
-
-func TestEvalMutation_RejectsInvalidPatchReturnTypes(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test"},
-		},
-	}
-	tests := []struct {
-		name      string
-		patchType string
-		expr      string
-	}{
-		{name: "apply configuration must return Object", patchType: "ApplyConfiguration", expr: "true"},
-		{name: "json patch must return list of JSONPatch", patchType: "JSONPatch", expr: "Object{}"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			policy := &AdmissionPolicy{
-				Mutations: []Mutation{
-					{
-						Path:       "spec.mutations[0]",
-						PatchType:  tt.patchType,
-						Expression: tt.expr,
-					},
-				},
-			}
-			_, err := e.EvalMutation(policy, input)
-			if err == nil {
-				t.Fatal("expected EvalMutation() to reject invalid patch return type")
-			}
-			if !strings.Contains(err.Error(), "spec.mutations[0]") {
-				t.Errorf("error should reference mutation path, got: %v", err)
-			}
-		})
-	}
-}
-
-func TestEvalMutation_WithoutPatchTypesRejectsPatchExpressions(t *testing.T) {
-	e, err := NewEvaluator(WithoutPatchTypes())
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Mutations: []Mutation{
-			{
-				Path:       "spec.mutations[0]",
-				PatchType:  "ApplyConfiguration",
-				Expression: "Object{}",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test"},
-		},
-	}
-
-	_, err = e.EvalMutation(policy, input)
-	if err == nil {
-		t.Fatal("expected EvalMutation() to reject patch expressions when patch types are disabled")
-	}
-}
-
-func TestEvalMutation_MultiplePatches(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Mutations: []Mutation{
-			{
-				Path:       "spec.mutations[0]",
-				PatchType:  "ApplyConfiguration",
-				Expression: "Object{metadata: Object.metadata{name: \"first\"}}",
-			},
-			{
-				Path:       "spec.mutations[1]",
-				PatchType:  "ApplyConfiguration",
-				Expression: "Object{metadata: Object.metadata{name: \"second\"}}",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "original"},
-		},
-	}
-
-	result, err := e.EvalMutation(policy, input)
-	if err != nil {
-		t.Fatalf("EvalMutation() error: %v", err)
-	}
-	if len(result.Patches) != 2 {
-		t.Fatalf("got %d patches, want 2", len(result.Patches))
-	}
-	wantNames := []string{"first", "second"}
-	for i, patch := range result.Patches {
-		if patch.Error != nil {
-			t.Errorf("patch[%d] error: %v", i, patch.Error)
-			continue
-		}
-		applyConfig, ok := patch.Value.(map[string]interface{})
-		if !ok {
-			t.Errorf("patch[%d] value type = %T, want map[string]interface{}", i, patch.Value)
-			continue
-		}
-		metadata, ok := applyConfig["metadata"].(map[string]interface{})
-		if !ok {
-			t.Errorf("patch[%d] metadata type = %T, want map[string]interface{}", i, applyConfig["metadata"])
-			continue
-		}
-		if metadata["name"] != wantNames[i] {
-			t.Errorf("patch[%d] metadata.name = %v, want %q", i, metadata["name"], wantNames[i])
-		}
-	}
-}
-
-func TestEvalMutation_CostLimitStopsFurtherMutations(t *testing.T) {
-	e, err := NewEvaluator(WithCostLimit(1))
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Mutations: []Mutation{
-			{
-				Path:       "spec.mutations[0]",
-				PatchType:  "ApplyConfiguration",
-				Expression: "Object{metadata: Object.metadata{name: object.metadata.name + object.metadata.name}}",
-			},
-			{
-				Path:       "spec.mutations[1]",
-				PatchType:  "ApplyConfiguration",
-				Expression: "Object{metadata: Object.metadata{name: \"second\"}}",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test-pod"},
-		},
-	}
-
-	result, err := e.EvalMutation(policy, input)
-	if err != nil {
-		t.Fatalf("EvalMutation() error: %v", err)
-	}
-	if len(result.Patches) != 1 {
-		t.Fatalf("got %d patches, want 1", len(result.Patches))
-	}
-	if result.Patches[0].Error == nil {
-		t.Fatal("expected first patch to fail after exceeding the cost budget")
-	}
-	if result.Cost < 0 || result.Cost > 1 {
-		t.Fatalf("Cost = %d, want between 0 and the configured limit", result.Cost)
-	}
-}
-
-func TestEvalMutationByPath(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Variables: []Variable{
-			{Name: "target", Expression: "'selected'"},
-		},
-		Mutations: []Mutation{
-			{
-				Path:       "spec.mutations[0]",
-				PatchType:  "ApplyConfiguration",
-				Expression: "Object{metadata: Object.metadata{name: \"first\"}}",
-			},
-			{
-				Path:       "spec.mutations[1]",
-				PatchType:  "ApplyConfiguration",
-				Expression: "Object{metadata: Object.metadata{name: variables.target}}",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "original"},
-		},
-	}
-
-	t.Run("select second mutation", func(t *testing.T) {
-		val, err := e.EvalMutationByPath(policy, MutationSelector{Path: "spec.mutations[1]"}, input)
-		if err != nil {
-			t.Fatalf("EvalMutationByPath() error: %v", err)
-		}
-		applyConfig, ok := val.(map[string]interface{})
-		if !ok {
-			t.Fatalf("value type = %T, want map[string]interface{}", val)
-		}
-		metadata, ok := applyConfig["metadata"].(map[string]interface{})
-		if !ok {
-			t.Fatalf("metadata type = %T, want map[string]interface{}", applyConfig["metadata"])
-		}
-		if metadata["name"] != "selected" {
-			t.Errorf("metadata.name = %v, want %q", metadata["name"], "selected")
-		}
-	})
-
-	t.Run("not found", func(t *testing.T) {
-		_, err := e.EvalMutationByPath(policy, MutationSelector{Path: "spec.mutations[99]"}, input)
-		if err == nil {
-			t.Error("expected error for missing path")
-		}
-	})
-
-	t.Run("empty path", func(t *testing.T) {
-		_, err := e.EvalMutationByPath(policy, MutationSelector{}, input)
-		if err == nil {
-			t.Error("expected error for empty path")
-		}
-	})
-}
-
-func TestParseAdmissionPolicyFile(t *testing.T) {
-	dir := t.TempDir()
-
-	t.Run("valid file", func(t *testing.T) {
-		path := filepath.Join(dir, "policy.yaml")
-		content := []byte(`
-variables:
-  - name: podName
-    expression: "object.metadata.name"
-validations:
-  - expression: "variables.podName != 'bad'"
-    message: "name must not be bad"
-`)
-		if err := os.WriteFile(path, content, 0644); err != nil {
-			t.Fatalf("writing test file: %v", err)
-		}
-		policy, err := ParseAdmissionPolicyFile(path)
-		if err != nil {
-			t.Fatalf("ParseAdmissionPolicyFile() error: %v", err)
-		}
-		if len(policy.Variables) != 1 {
-			t.Errorf("got %d variables, want 1", len(policy.Variables))
-		}
-		if len(policy.Validations) != 1 {
-			t.Errorf("got %d validations, want 1", len(policy.Validations))
-		}
-	})
-
-	t.Run("missing file", func(t *testing.T) {
-		_, err := ParseAdmissionPolicyFile(filepath.Join(dir, "nonexistent.yaml"))
-		if err == nil {
-			t.Error("expected error for missing file")
-		}
-	})
-
-	t.Run("invalid content", func(t *testing.T) {
-		path := filepath.Join(dir, "empty.yaml")
-		if err := os.WriteFile(path, []byte("{}"), 0644); err != nil {
-			t.Fatalf("writing test file: %v", err)
-		}
-		_, err := ParseAdmissionPolicyFile(path)
-		if err == nil {
-			t.Error("expected error for empty policy file")
 		}
 	})
 }
@@ -1672,324 +206,6 @@ func TestWithVersion(t *testing.T) {
 	if err != nil {
 		t.Errorf("CompileCheck with sets at v1.29 should succeed, got: %v", err)
 	}
-}
-
-func TestParseAdmissionInput(t *testing.T) {
-	yaml := `
-object:
-  apiVersion: v1
-  kind: Pod
-  metadata:
-    name: test-pod
-    namespace: default
-params:
-  apiVersion: v1
-  kind: ConfigMap
-  metadata:
-    name: policy-config
-  data:
-    allowed: "true"
-request:
-  operation: CREATE
-  kind:
-    version: v1
-    kind: Pod
-  resource:
-    version: v1
-    resource: pods
-namespaceObject:
-  apiVersion: v1
-  kind: Namespace
-  metadata:
-    name: default
-`
-	input, err := ParseAdmissionInput(yaml)
-	if err != nil {
-		t.Fatalf("ParseAdmissionInput() error: %v", err)
-	}
-	if input.Object["kind"] != "Pod" {
-		t.Errorf("object.kind = %v, want Pod", input.Object["kind"])
-	}
-	if input.Params["kind"] != "ConfigMap" {
-		t.Errorf("params.kind = %v, want ConfigMap", input.Params["kind"])
-	}
-	if input.Request == nil || input.Request.Operation != admissionv1.Create {
-		t.Fatalf("request.operation = %v, want CREATE", input.Request)
-	}
-	if input.Namespace == nil || input.Namespace.Name != "default" {
-		t.Fatalf("namespace.name = %v, want default", input.Namespace)
-	}
-
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-	policy := &AdmissionPolicy{
-		Validations: []Validation{{Path: "validations[0]", Expression: "params.data.allowed == 'true'"}},
-	}
-	policy.SetHasParams(true)
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if !result.Allowed {
-		t.Errorf("expected Allowed=true, got violations: %s", result.FormatViolations())
-	}
-}
-
-func TestParseAdmissionInput_PreambleCanUnwrapParamsResource(t *testing.T) {
-	yaml := `
-object:
-  apiVersion: v1
-  kind: Pod
-  metadata:
-    name: labeled-pod
-    labels:
-      team: platform
-params:
-  apiVersion: policy.example.com/v1alpha1
-  kind: LabelPolicyParams
-  metadata:
-    name: required-labels
-  spec:
-    parameters:
-      labels:
-      - key: team
-request:
-  operation: CREATE
-  kind:
-    version: v1
-    kind: Pod
-  resource:
-    version: v1
-    resource: pods
-`
-	input, err := ParseAdmissionInput(yaml)
-	if err != nil {
-		t.Fatalf("ParseAdmissionInput() error: %v", err)
-	}
-
-	e, err := NewEvaluator(WithPreambleVariables(
-		Variable{
-			Name:       "params",
-			Expression: `!has(params.spec) ? null : !has(params.spec.parameters) ? null : params.spec.parameters`,
-		},
-	))
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{
-				Path:       "validations[0]",
-				Expression: `variables.params.labels.all(entry, has(object.metadata.labels) && entry.key in object.metadata.labels)`,
-				Message:    "missing required label",
-			},
-		},
-	}
-	policy.SetHasParams(true)
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if !result.Allowed {
-		t.Errorf("expected Allowed=true, got violations: %s", result.FormatViolations())
-	}
-}
-
-func TestSetHasParams(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	t.Run("params enabled", func(t *testing.T) {
-		policy := &AdmissionPolicy{
-			Validations: []Validation{
-				{Path: "validations[0]", Expression: "params.data.maxReplicas == '5'"},
-			},
-		}
-		policy.SetHasParams(true)
-
-		input := &AdmissionInput{
-			Object: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "test"},
-			},
-			Params: map[string]interface{}{
-				"data": map[string]interface{}{"maxReplicas": "5"},
-			},
-		}
-
-		result, err := e.EvalAdmission(policy, input)
-		if err != nil {
-			t.Fatalf("EvalAdmission() error: %v", err)
-		}
-		if !result.Allowed {
-			t.Errorf("expected Allowed=true, got violations: %s", result.FormatViolations())
-		}
-	})
-
-	t.Run("params disabled causes compilation error", func(t *testing.T) {
-		policy := &AdmissionPolicy{
-			Validations: []Validation{
-				{Path: "validations[0]", Expression: "params.data.maxReplicas == '5'"},
-			},
-		}
-		policy.SetHasParams(false)
-
-		input := &AdmissionInput{
-			Object: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "test"},
-			},
-		}
-
-		_, err := e.EvalAdmission(policy, input)
-		if err == nil {
-			t.Error("expected compilation error when params is disabled but expression references params")
-		}
-	})
-
-	t.Run("default enables params", func(t *testing.T) {
-		// Manually constructed policy without SetHasParams should default to params enabled.
-		policy := &AdmissionPolicy{
-			Validations: []Validation{
-				{Path: "validations[0]", Expression: "params.data.key == 'val'"},
-			},
-		}
-		input := &AdmissionInput{
-			Object: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "test"},
-			},
-			Params: map[string]interface{}{
-				"data": map[string]interface{}{"key": "val"},
-			},
-		}
-
-		result, err := e.EvalAdmission(policy, input)
-		if err != nil {
-			t.Fatalf("EvalAdmission() error: %v", err)
-		}
-		if !result.Allowed {
-			t.Errorf("expected Allowed=true with default params, got violations: %s", result.FormatViolations())
-		}
-	})
-}
-
-func TestEvalAdmission_OldObject(t *testing.T) {
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Validations: []Validation{
-			{
-				Path:       "validations[0]",
-				Expression: "object.metadata.name != oldObject.metadata.name",
-				Message:    "name must change on update",
-			},
-		},
-	}
-
-	t.Run("names differ", func(t *testing.T) {
-		input := &AdmissionInput{
-			Object: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "new-name"},
-			},
-			OldObject: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "old-name"},
-			},
-			Request: &admissionv1.AdmissionRequest{
-				Operation: admissionv1.Update,
-				Kind:      metav1.GroupVersionKind{Version: "v1", Kind: "Pod"},
-				Resource:  metav1.GroupVersionResource{Version: "v1", Resource: "pods"},
-			},
-		}
-
-		result, err := e.EvalAdmission(policy, input)
-		if err != nil {
-			t.Fatalf("EvalAdmission() error: %v", err)
-		}
-		if !result.Allowed {
-			t.Errorf("expected Allowed=true, got violations: %s", result.FormatViolations())
-		}
-	})
-
-	t.Run("names same", func(t *testing.T) {
-		input := &AdmissionInput{
-			Object: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "same-name"},
-			},
-			OldObject: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "same-name"},
-			},
-			Request: &admissionv1.AdmissionRequest{
-				Operation: admissionv1.Update,
-				Kind:      metav1.GroupVersionKind{Version: "v1", Kind: "Pod"},
-				Resource:  metav1.GroupVersionResource{Version: "v1", Resource: "pods"},
-			},
-		}
-
-		result, err := e.EvalAdmission(policy, input)
-		if err != nil {
-			t.Fatalf("EvalAdmission() error: %v", err)
-		}
-		if result.Allowed {
-			t.Error("expected Allowed=false when names are the same")
-		}
-		if len(result.Violations) != 1 {
-			t.Fatalf("got %d violations, want 1", len(result.Violations))
-		}
-		if result.Violations[0].Message != "name must change on update" {
-			t.Errorf("violation message = %q, want %q", result.Violations[0].Message, "name must change on update")
-		}
-	})
-
-	t.Run("inferred update operation", func(t *testing.T) {
-		// When both object and oldObject are set without explicit request, operation should be inferred as UPDATE.
-		input := &AdmissionInput{
-			Object: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "new-name"},
-			},
-			OldObject: map[string]interface{}{
-				"apiVersion": "v1",
-				"kind":       "Pod",
-				"metadata":   map[string]interface{}{"name": "old-name"},
-			},
-		}
-
-		policyOp := &AdmissionPolicy{
-			Validations: []Validation{
-				{Path: "validations[0]", Expression: "request.operation == 'UPDATE'"},
-			},
-		}
-
-		result, err := e.EvalAdmission(policyOp, input)
-		if err != nil {
-			t.Fatalf("EvalAdmission() error: %v", err)
-		}
-		if !result.Allowed {
-			t.Errorf("expected inferred UPDATE operation, got violations: %s", result.FormatViolations())
-		}
-	})
 }
 
 func TestWithCostLimit(t *testing.T) {
@@ -2030,218 +246,5 @@ func TestWithCostLimit(t *testing.T) {
 		if result.Cost <= 0 {
 			t.Error("expected Cost > 0 even with low budget")
 		}
-	}
-}
-
-func TestParseAdmissionPolicy_ValidatingWebhookConfiguration(t *testing.T) {
-	yaml := `
-apiVersion: admissionregistration.k8s.io/v1
-kind: ValidatingWebhookConfiguration
-webhooks:
-  - name: validate.example.com
-    matchConditions:
-      - name: exclude-leases
-        expression: "!(request.resource.resource == 'leases')"
-      - name: exclude-system-ns
-        expression: "request.namespace != 'kube-system'"
-`
-	policy, err := ParseAdmissionPolicy(yaml)
-	if err != nil {
-		t.Fatalf("ParseAdmissionPolicy() error: %v", err)
-	}
-	if len(policy.MatchConditions) != 2 {
-		t.Fatalf("got %d matchConditions, want 2", len(policy.MatchConditions))
-	}
-	if policy.MatchConditions[0].Path != "webhooks[0].matchConditions[0]" {
-		t.Errorf("matchCondition[0] path = %q, want %q", policy.MatchConditions[0].Path, "webhooks[0].matchConditions[0]")
-	}
-	if policy.hasParams {
-		t.Error("hasParams should be false for webhook configurations")
-	}
-}
-
-func TestParseAdmissionPolicy_MutatingWebhookConfiguration(t *testing.T) {
-	yaml := `
-apiVersion: admissionregistration.k8s.io/v1
-kind: MutatingWebhookConfiguration
-webhooks:
-  - name: mutate.example.com
-    matchConditions:
-      - name: only-pods
-        expression: "request.resource.resource == 'pods'"
-`
-	policy, err := ParseAdmissionPolicy(yaml)
-	if err != nil {
-		t.Fatalf("ParseAdmissionPolicy() error: %v", err)
-	}
-	if len(policy.MatchConditions) != 1 {
-		t.Fatalf("got %d matchConditions, want 1", len(policy.MatchConditions))
-	}
-	if policy.MatchConditions[0].Expression != "request.resource.resource == 'pods'" {
-		t.Errorf("unexpected expression: %q", policy.MatchConditions[0].Expression)
-	}
-}
-
-func TestParseAdmissionPolicy_WebhookErrors(t *testing.T) {
-	tests := []struct {
-		name string
-		yaml string
-	}{
-		{name: "empty matchConditions", yaml: `
-apiVersion: admissionregistration.k8s.io/v1
-kind: ValidatingWebhookConfiguration
-webhooks:
-  - name: validate.example.com
-`},
-		{name: "empty expression in matchCondition", yaml: `
-apiVersion: admissionregistration.k8s.io/v1
-kind: ValidatingWebhookConfiguration
-webhooks:
-  - name: validate.example.com
-    matchConditions:
-      - name: empty
-        expression: ""
-`},
-		{name: "mutating empty matchConditions", yaml: `
-apiVersion: admissionregistration.k8s.io/v1
-kind: MutatingWebhookConfiguration
-webhooks:
-  - name: mutate.example.com
-`},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, err := ParseAdmissionPolicy(tt.yaml)
-			if err == nil {
-				t.Error("ParseAdmissionPolicy() expected error")
-			}
-		})
-	}
-}
-
-func TestEvalAdmission_WebhookMatchConditions(t *testing.T) {
-	e, err := NewEvaluator(WithoutAuthorizer(), WithoutPatchTypes())
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	yaml := `
-apiVersion: admissionregistration.k8s.io/v1
-kind: ValidatingWebhookConfiguration
-webhooks:
-  - name: validate.example.com
-    matchConditions:
-      - name: only-pods
-        expression: "request.resource.resource == 'pods'"
-`
-	policy, err := ParseAdmissionPolicy(yaml)
-	if err != nil {
-		t.Fatalf("ParseAdmissionPolicy() error: %v", err)
-	}
-
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test"},
-		},
-		Request: &admissionv1.AdmissionRequest{
-			Kind:     metav1.GroupVersionKind{Version: "v1", Kind: "Pod"},
-			Resource: metav1.GroupVersionResource{Version: "v1", Resource: "pods"},
-		},
-	}
-
-	result, err := e.EvalAdmission(policy, input)
-	if err != nil {
-		t.Fatalf("EvalAdmission() error: %v", err)
-	}
-	if !result.Allowed {
-		t.Errorf("expected Allowed=true for pods, got violations: %s", result.FormatViolations())
-	}
-}
-
-func TestDefaultResourceForGVK(t *testing.T) {
-	tests := []struct {
-		kind    string
-		wantRes string
-	}{
-		{kind: "Pod", wantRes: "pods"},
-		{kind: "Service", wantRes: "services"},
-		{kind: "Deployment", wantRes: "deployments"},
-		{kind: "Ingress", wantRes: "ingresses"},
-		{kind: "NetworkPolicy", wantRes: "networkpolicies"},
-		{kind: "DaemonSet", wantRes: "daemonsets"},
-		{kind: "ConfigMap", wantRes: "configmaps"},
-		{kind: "Endpoints", wantRes: "endpointses"}, // naive but consistent for sibilant ending
-		{kind: "", wantRes: "objects"},
-	}
-	for _, tt := range tests {
-		t.Run(tt.kind, func(t *testing.T) {
-			gvk := schema.GroupVersionKind{Version: "v1", Kind: tt.kind}
-			gvr := defaultResourceForGVK(gvk)
-			if gvr.Resource != tt.wantRes {
-				t.Errorf("defaultResourceForGVK(%q) = %q, want %q", tt.kind, gvr.Resource, tt.wantRes)
-			}
-		})
-	}
-}
-
-func TestEvalMutation_CompilationError(t *testing.T) {
-	// Verify that EvalMutation returns an error (not a result with per-patch errors)
-	// for compilation failures, symmetric with EvalAdmission.
-	e, err := NewEvaluator()
-	if err != nil {
-		t.Fatalf("NewEvaluator() error: %v", err)
-	}
-
-	policy := &AdmissionPolicy{
-		Mutations: []Mutation{
-			{
-				Path:       "spec.mutations[0]",
-				PatchType:  "ApplyConfiguration",
-				Expression: "Object{metadata: Object.metadata{name: 'valid'}}",
-			},
-			{
-				Path:       "spec.mutations[1]",
-				PatchType:  "ApplyConfiguration",
-				Expression: "???invalid",
-			},
-		},
-	}
-	input := &AdmissionInput{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata":   map[string]interface{}{"name": "test"},
-		},
-	}
-
-	_, err = e.EvalMutation(policy, input)
-	if err == nil {
-		t.Error("EvalMutation() expected error when any mutation has compilation error")
-	}
-	if !strings.Contains(err.Error(), "spec.mutations[1]") {
-		t.Errorf("error should reference the failing mutation path, got: %v", err)
-	}
-}
-
-func TestParseAdmissionPolicy_VAPWithParamKind(t *testing.T) {
-	yaml := `
-apiVersion: admissionregistration.k8s.io/v1
-kind: ValidatingAdmissionPolicy
-spec:
-  paramKind:
-    apiVersion: v1
-    kind: ConfigMap
-  validations:
-    - expression: "params.data.allowed == 'true'"
-      message: "not allowed"
-`
-	policy, err := ParseAdmissionPolicy(yaml)
-	if err != nil {
-		t.Fatalf("ParseAdmissionPolicy() error: %v", err)
-	}
-	if !policy.hasParams {
-		t.Error("hasParams should be true when paramKind is set")
 	}
 }
